Update islemGames.py
divide opp. does not have to integer.
now it has float operations so it find more effectively.

+++ gamesAlgorithms/islemGames.py
@@ -21,7 +21,7 @@ def carp(x,y):
 def bol(x,y):
     if y != 0:
         #bol =np.float(x)/y
-        bol = x/y
+        bol = np.float(x)/y
     else:
         bol = 0
     return bol
